Refactored main CLI package location
